签到奖励显示INT

+++ app/src/main/java/org/cxct/sportlottery/net/money/data/FirstDepositDetail.kt
@@ -56,7 +56,7 @@ data class FirstDepositDetail(
     /**
      * 次日奖励金额
      */
-    val rewardAmount: Double?
+    val rewardAmount: Int?
 
 ):Parcelable{
     fun getCurrentDepositConfig():FirstDepositConfig? =
